requirements.txt
flask
numpy
matplotlib
requests
openai
pytest
markdown2
python-dotenv
scipy
google-generativeai 